move srvplot to components
